+++ pkg/genericapiserver/registry/rest/BUILD
@@ -21,9 +21,9 @@ go_library(
     ],
     tags = ["automanaged"],
     deps = [
-        "//pkg/api/validation/genericvalidation:go_default_library",
         "//vendor:k8s.io/apimachinery/pkg/api/errors",
         "//vendor:k8s.io/apimachinery/pkg/api/meta",
+        "//vendor:k8s.io/apimachinery/pkg/api/validation",
         "//vendor:k8s.io/apimachinery/pkg/api/validation/path",
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/internalversion",
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/v1",
